@@ -13,16 +13,16 @@
   "private": true,
   "dependencies": {
     "@amcharts/amcharts4": "^4.9.13",
-    "@angular/animations": "^13.4.0",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/fire": "^7.6.1",
-    "@angular/forms": "^13.4.0",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/platform-server": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/platform-server": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@fortawesome/fontawesome-free": "^5.5.0",
     "@techiediaries/ngx-qrcode": "^9.1.0",
     "@types/chartist": "^0.9.43",
@@ -62,11 +62,11 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1303.11",
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular/cli": "^13.3.11",
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/language-service": "^13.4.0",
+    "@angular-devkit/architect": "0.1402.13",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "^3.0.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^10.12.9",
