@@ -5,7 +5,7 @@ import { map } from 'rxjs/operators';
 import { SummarizeService } from 'src/app/shared/services/summarize.service';
 import { Subscription } from 'rxjs';
 import { NzMessageService, NzModalService } from 'ng-zorro-antd';
-import { FormBuilder, FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, FormControl, Validators } from '@angular/forms';
 import { RolService } from 'src/app/shared/services/roles.service';
 
 @Component({
@@ -26,7 +26,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
   isVisible = false;
   isDeleteVisible = false;
   isOkLoading = false;
-  validateForm: FormGroup;
+  validateForm: UntypedFormGroup;
   infoLoad: any = [];
   userlevelAccess:string;
 
@@ -37,7 +37,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
     private messageService: NzMessageService,
     public modalService: NzModalService,
     public rolService: RolService,
-    private fb: FormBuilder
+    private fb: UntypedFormBuilder
     ) { }
 
   ngOnInit() {
