@@ -6,7 +6,7 @@ import { Subject } from 'rxjs';
 import { takeUntil, map } from 'rxjs/operators';
 import { IRoute } from 'src/app/shared/interfaces/route.type';
 import { AccountsService } from 'src/app/shared/services/accounts.service';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, Validators } from '@angular/forms';
 import * as _ from 'lodash';
 import { RolService } from 'src/app/shared/services/roles.service';
 
@@ -24,7 +24,7 @@ export class RoutesComponent implements OnInit, OnDestroy {
   stopSubscription$: Subject<boolean> = new Subject();
   routesList: any = [];
   accountsList: any = [];
-  objectForm: FormGroup;
+  objectForm: UntypedFormGroup;
 
   confirmModal: NzModalRef;
   duplicateVisible: boolean = false;
@@ -42,7 +42,7 @@ export class RoutesComponent implements OnInit, OnDestroy {
     private accountsService: AccountsService,
     private messageService: NzMessageService,
     private rolService: RolService,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private modal: NzModalService
   ) { }
 
